requirements.txt
Flask==2.1.3
requests==2.28.1
psycopg2-binary
python-dotenv